Rename upload to localUpload in Redux state
