+++ public/video-ui/src/reducers/localUploadsReducer.js
@@ -1,4 +1,4 @@
-export default function upload(state = { handle: null, progress: 0, total: 0 }, action) {
+export default function localUpload(state = { handle: null, progress: 0, total: 0 }, action) {
   switch(action.type) {
     case 'UPLOAD_STARTED': {
       const total = action.upload.parts[action.upload.parts.length - 1].end;
